+++ packages/core/src/utils/index.ts
@@ -506,30 +506,34 @@ export function checkIsCol(str: string) {
   return null;
 }
 
-export function getLineCount(sentence: string, maxWidthPx: number, font = '12px Arial') {
-  const canvas = document.createElement('canvas');
-  const context = canvas.getContext('2d');
+export function getLineCount(
+  sentence: string,
+  maxWidthPx: number,
+  font = "12px Arial"
+) {
+  const canvas = document.createElement("canvas");
+  const context = canvas.getContext("2d");
   if (context) {
     context.font = font;
   }
 
-  const words = sentence.split(' ');
-  let currentLine = '';
+  const words = sentence.split(" ");
+  let currentLine = "";
   let lineCount = 0;
 
-  words.forEach(word => {
-    const testLine = currentLine ? currentLine + ' ' + word : word;
+  words.forEach((word) => {
+    const testLine = currentLine ? `${currentLine} ${word}` : word;
     const testWidth = context?.measureText(testLine).width;
 
     if (testWidth && testWidth > maxWidthPx && currentLine) {
-      lineCount++;
+      lineCount += 1;
       currentLine = word;
     } else {
       currentLine = testLine;
     }
   });
 
-  if (currentLine) lineCount++;
+  if (currentLine) lineCount += 1;
 
   return lineCount;
 }