+++ packages/react/src/components/Toolbar/index.tsx
@@ -1786,14 +1786,19 @@ const Toolbar: React.FC<{
                           value,
                           refs.canvas.current!.getContext("2d")!
                         );
-                        const sheetIndex = getSheetIndex(ctx, ctx.currentSheetId) || 0;
+                        const sheetIndex =
+                          getSheetIndex(ctx, ctx.currentSheetId) || 0;
                         const config = ctx.luckysheetfile[sheetIndex].config!;
-                        const currentRowHeight = config?.rowlen?.[0] ||
-                                ctx.luckysheetfile[sheetIndex].defaultColWidth ||
-                                  22;
-                        api.setRowHeight(ctx, { 1: currentRowHeight+0.2 }, { id: ctx.currentSheetId });
-
-                      }); 
+                        const currentRowHeight =
+                          config?.rowlen?.[0] ||
+                          ctx.luckysheetfile[sheetIndex].defaultColWidth ||
+                          22;
+                        api.setRowHeight(
+                          ctx,
+                          { 1: currentRowHeight + 0.2 },
+                          { id: ctx.currentSheetId }
+                        );
+                      });
                       setOpen(false);
                     }}
                     tabIndex={0}