+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @fileverse-dev/fortune-core
 
+## 1.2.30
+
+### Patch Changes
+
+- wrap height oggset
+
 ## 1.2.29
 
 ### Patch Changes